grpc restructure for the plant

+++ Entities/plant/Watering.cs
@@ -8,7 +8,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 public class Watering
 {
     [Key] public int Id {get; set;}
-    public double PumpTimeInSeconds { get; set; } = 0;
+    public int PumpTimeInSeconds { get; set; } = 0;
     public DateTime LastWaterTime { get; set; }
     public DateTime PredictedFutureWaterTime { get; set; }
